### 2016-10-18 #96 @whedon
[PRE REVIEW]: popeye: a population receptive field estimation tool
**Submitting author:** @kdesimone (<a href="http://orcid.org/0000-0003-4254-9408">Kevin DeSimone</a>)
**Repository:** <a href="https://github.com/kdesimone/popeye" target ="_blank">https://github.com/kdesimone/popeye</a>
**Version:** v.1.0.0
