### 2016-10-20 #96 @arokem
Maybe someone  from the psychopy community would be willing and able to do this?

cc @psychopy/psychopy
