### 2016-10-21 #96 @arokem
Another direction are members of the Nilearn team (@bthirion and his team, for example)
