### 2016-10-25 #96 @kdesimone
I was going to suggest nipy contributors ...

https://github.com/orgs/nipy/people

On Thu, Oct 20, 2016 at 3:11 PM, Ariel Rokem <notifications@github.com>
wrote:

> Maybe someone from the psychopy community would be willing and able to do
> this?
>
> cc @psychopy/psychopy
>
