### 2016-10-28 057e30d
RF: fixed up the manual grid slicer to be inclusive of the upper and lowr bounds
