### 2016-10-28 ea31231
TST: updated test coverage
