### 2016-10-28 #96 @cMadan
@tknapen - awesome, thank you! Please see the reviewer guidelines here http://joss.theoj.org/about#reviewer_guidelines
