### 2016-10-28 #103 @cMadan
@tknapen - you should be able to check the boxes in the first post in this issue. For some additional details, please see the reviewer guidelines (http://joss.theoj.org/about#reviewer_guidelines). Otherwise, let me know if you have any questions!
