### 2016-10-28 3bbcb13
TST: extended coverage for HRF without the integrator
