### 2016-10-28 37843bc
RF: ignoring some switches for coverage report
