### 2016-10-28 bb1a75f
ENH: adding markdown README
