### 2016-10-28 63dc17f
RF: removing newlines
