### 2016-10-28 29690ae
RM: removed the rst version of the readme
