### 2016-10-28 3e959f6
RF: updated the dependencies
