### 2016-10-29 4919355
RF: changed naming convention for the coarse approximation fix
