### 2016-10-29 06c795b
RF: changed naming convention for the coarse approximation fix
