### 2016-11-08 efed709
RF: updated call signature of auditory model
