### 2016-11-16 952d4a0
ENH: added new function for extracting M/P responses
