### 2016-11-16 4821013
ENH: added new spatiotemporal model using M/P betas instead of weight
