### 2016-11-16 b21ad42
ENH: added new spatiotemporal model using M/P betas instead of weight and css
