+++ popeye/spatiotemporal_css_betas_masked.py
@@ -1,0 +1,312 @@
+#!/usr/bin/python
+
+""" Classes and functions for fitting SpatioTemporal population receptive field models """
+
+from __future__ import division, print_function, absolute_import
+import time
+import gc
+import warnings
+warnings.simplefilter("ignore")
+
+import numpy as np
+from scipy.stats import linregress
+from scipy.signal import fftconvolve
+from scipy.integrate import trapz, simps
+import nibabel
+
+from popeye.onetime import auto_attr
+import popeye.utilities as utils
+from popeye.base import PopulationModel, PopulationFit
+from popeye.spinach import generate_og_receptive_field, generate_strf_timeseries, generate_rf_timeseries
+
+class SpatioTemporalModel(PopulationModel):
+    
+    """
+    A Gaussian population receptive field model class
+    
+    """
+    
+    def __init__(self, stimulus, hrf_model):
+        
+        """
+        A spatiotemporal population receptive field model.
+        
+        Paramaters
+        ----------
+        
+        stimulus : `VisualStimulus` class object
+            A class instantiation of the `VisualStimulus` class
+            containing a representation of the visual stimulus.
+        
+        
+        """
+        
+        PopulationModel.__init__(self, stimulus, hrf_model)
+    
+    
+    # for the final solution, we use spatiotemporal
+    def generate_ballpark_prediction(self, x, y, sigma, n, mbeta, pbeta, baseline):
+        
+        # mask for speed
+        mask = self.distance_mask_coarse(x, y, sigma*3)
+        
+        # generate the RF
+        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
+        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
+        
+        # spatial_response
+        rf_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
+        
+        # compression
+        rf_ts **= n
+        
+        # temporal response
+        m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
+        
+        # # normalize each timeseries
+        # m_ts = utils.normalize(m_ts,0,1)
+        # p_ts = utils.normalize(p_ts,0,1)
+        
+        # convolve with HRF
+        hrf = self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
+        
+        # M
+        m_model = fftconvolve(m_ts, hrf)[0:len(m_ts)]
+        m_model /= len(m_model)
+        
+        # P
+        p_model = fftconvolve(p_ts, hrf)[0:len(p_ts)]
+        p_model /= len(p_model)
+        
+        # mix
+        model = m_model * mbeta + p_model * pbeta
+         
+        # offset
+        model += baseline
+        
+        return model
+    
+    # for the final solution, we use spatiotemporal
+    def generate_prediction(self, x, y, sigma, n, mbeta, pbeta, baseline):
+        
+        # mask for speed
+        mask = self.distance_mask(x, y, sigma*3)
+        
+        # generate the RF
+        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x, self.stimulus.deg_y)
+        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x[0,0:2])**2)
+        
[…]
