### 2016-11-16 837edbb
RF: updated new M/P responses
