+++ popeye/spatiotemporal_css_betas_masked.py
@@ -17,7 +17,7 @@ import nibabel
 from popeye.onetime import auto_attr
 import popeye.utilities as utils
 from popeye.base import PopulationModel, PopulationFit
-from popeye.spinach import generate_og_receptive_field, generate_strf_timeseries, generate_rf_timeseries
+from popeye.spinach import generate_og_receptive_field, generate_mp_timeseries, generate_rf_timeseries
 
 class SpatioTemporalModel(PopulationModel):
     
@@ -48,35 +48,29 @@ class SpatioTemporalModel(PopulationModel):
     def generate_ballpark_prediction(self, x, y, sigma, n, mbeta, pbeta, baseline):
         
         # mask for speed
-        mask = self.distance_mask_coarse(x, y, sigma*3)
+        mask = self.distance_mask_coarse(x, y, sigma*self.mask_size)
         
         # generate the RF
         spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
         spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
         
         # spatial_response
-        rf_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
+        spatial_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
         
         # compression
-        rf_ts **= n
+        spatial_ts **= n
         
         # temporal response
-        m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
-        
-        # # normalize each timeseries
-        # m_ts = utils.normalize(m_ts,0,1)
-        # p_ts = utils.normalize(p_ts,0,1)
+        m_ts, p_ts = generate_mp_timeseries(spatial_ts, self.m_amp, self.p_amp, self.stimulus.flicker_vec)
         
         # convolve with HRF
         hrf = self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
         
         # M
         m_model = fftconvolve(m_ts, hrf)[0:len(m_ts)]
-        m_model /= len(m_model)
         
         # P
         p_model = fftconvolve(p_ts, hrf)[0:len(p_ts)]
-        p_model /= len(p_model)
         
         # mix
         model = m_model * mbeta + p_model * pbeta
@@ -90,39 +84,29 @@ class SpatioTemporalModel(PopulationModel):
     def generate_prediction(self, x, y, sigma, n, mbeta, pbeta, baseline):
         
         # mask for speed
-        mask = self.distance_mask(x, y, sigma*3)
+        mask = self.distance_mask_coarse(x, y, sigma*self.mask_size)
         
         # generate the RF
-        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x, self.stimulus.deg_y)
-        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x[0,0:2])**2)
+        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
+        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
         
-        # spatial response
-        rf_ts = generate_rf_timeseries(self.stimulus.stim_arr, spatial_rf, mask)
+        # spatial_response
[…]
