### 2016-11-17 d03c157
RF: updated the M/P timeseries generation
