+++ popeye/spatiotemporal_masked.py
@@ -17,7 +17,7 @@ import nibabel
 from popeye.onetime import auto_attr
 import popeye.utilities as utils
 from popeye.base import PopulationModel, PopulationFit
-from popeye.spinach import generate_og_receptive_field, generate_strf_timeseries, generate_rf_timeseries
+from popeye.spinach import generate_og_receptive_field, generate_mp_timeseries, generate_rf_timeseries
 
 class SpatioTemporalModel(PopulationModel):
     
@@ -55,18 +55,10 @@ class SpatioTemporalModel(PopulationModel):
         spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
         
         # spatial_response
-        rf_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
+        spatial_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
         
         # temporal response
-        m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
-        
-        # clean up nan
-        m_ts[np.isnan(m_ts)] = 0
-        p_ts[np.isnan(p_ts)] = 0
-        
-        # normalize each timeseries
-        m_ts = utils.normalize(m_ts,0,1)
-        p_ts = utils.normalize(p_ts,0,1)
+        m_ts, p_ts = generate_mp_timeseries(spatial_ts, self.m_amp, self.p_amp, self.stimulus.flicker_vec)
         
         # mix them
         mp_ts = (1-weight) * m_ts + weight * p_ts
@@ -94,18 +86,10 @@ class SpatioTemporalModel(PopulationModel):
         spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x[0,0:2])**2)
         
         # spatial response
-        rf_ts = generate_rf_timeseries(self.stimulus.stim_arr, spatial_rf, mask)
+        spatial_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
         
         # temporal response
-        m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
-        
-        # clean up nan
-        m_ts[np.isnan(m_ts)] = 0
-        p_ts[np.isnan(p_ts)] = 0
-        
-        # normalize each timeseries
-        m_ts = utils.normalize(m_ts,0,1)
-        p_ts = utils.normalize(p_ts,0,1)
+        m_ts, p_ts = generate_mp_timeseries(spatial_ts, self.m_amp, self.p_amp, self.stimulus.flicker_vec)
         
         # mix them
         mp_ts = (1-weight) * m_ts + weight * p_ts 
@@ -155,40 +139,27 @@ class SpatioTemporalModel(PopulationModel):
     
     @auto_attr
     def flickers(self):
-        ts = self.t.shape[-1]
-        fs = int(np.max(self.stimulus.flicker_vec))
-        flickers = np.zeros((ts,fs))
-        for f in xrange(1,fs+1):
-            flickers[:,f-1] = np.sin(2 * np.pi * self.stimulus.flicker_hz[f-1] * self.t)
-        return flickers
+        return np.sin(2 * np.pi * np.single(self.stimulus.flicker_hz) * self.t[:,np.newaxis])
     
     @auto_attr
     def p_resp(self):
-        ts = self.t.shape[-1]
-        fs = int(np.max(self.stimulus.flicker_vec))
[…]
