### 2016-11-17 8d04ddd
RF: mask_size is now passed into model by user
