### 2016-11-17 ec39f3e
ENH: added new nonsticky model methods
