### 2016-11-22 959f18f
RF: taking the sum of the rectified M/P response
