### 2016-11-22 bf1a780
RF: taking the sum of the rectified M/P response
