### 2016-11-22 72ba1a5
ENH: added new version of the model with pct sig ch
