### 2016-11-22 cdb76b2
Python 3 compatibility.
