+++ popeye/tests/test_utilities.py
@@ -1,6 +1,10 @@
 from __future__ import division
 import ctypes, sharedmem, sys
-from StringIO import StringIO
+
+try:
+    from StringIO import StringIO
+except ImportError:
+    from io import StringIO
 
 import numpy as np
 import nose.tools as nt
@@ -15,7 +19,7 @@ import popeye.og as og
 from popeye.visual_stimulus import VisualStimulus, simulate_bar_stimulus, resample_stimulus, generate_coordinate_matrices
 
 def test_distance_mask():
-    
+
     x = 0
     y = 0
     sigma = 1
@@ -27,82 +31,82 @@ def test_distance_mask():
     mask = utils.distance_mask(x,y,sigma,dx,dy,amplitude)
     nt.assert_true(np.max(mask)==amplitude)
 
-    
+
 def test_grid_slice():
-    
+
     # test this case
     from_1 = 0
     to_1 = 20
     from_2 = 0
     to_2 = 2
     Ns=5
-    
+
     # set a parameter to estimate
     params = (10,1)
-    
+
     # see if we properly tile the parameter space for Ns=2
     grid_1 = utils.grid_slice(from_1, to_1, Ns)
     grid_2 = utils.grid_slice(from_2, to_2, Ns)
     grids = (grid_1, grid_2)
-    
+
     # unbounded
     bounds = ()
-    
+
     # create a simple function to generate a response from the parameter
     func = lambda freq,offset: np.sin( np.linspace(0,1,1000) * 2 * np.pi * freq) + offset
-    
+
     # create a "response"
     response = func(*params)
-    
+
     # get the ball-park estimate
     p0 = utils.brute_force_search(response, utils.error_function, func, grids, bounds)
-    
+
     # make sure we fit right
     npt.assert_equal(params, p0[0])
-    
+
     # make sure we sliced it right
     npt.assert_equal(p0[2][0].min(),from_1)
     npt.assert_equal(p0[2][0].max(),to_1)
     npt.assert_equal(p0[2][1].min(),from_2)
     npt.assert_equal(p0[2][1].max(),to_2)
-    
+
     # test this case
     from_1 = 0
     to_1 = 20
     from_2 = 1
     to_2 = 2
     Ns=2
-    
+
     # set a parameter to estimate
     params = (0,1)
-    
+
     # see if we properly tile the parameter space for Ns=2
     grid_1 = utils.grid_slice(from_1, to_1, Ns)
     grid_2 = utils.grid_slice(from_2, to_2, Ns)
     grids = (grid_1, grid_2)
-    
+
     # unbounded
     bounds = ()
-    
+
     # create a simple function to generate a response from the parameter
     func = lambda freq,offset: np.sin( np.linspace(0,1,1000) * 2 * np.pi * freq) + offset
-    
+
     # create a "response"
     response = func(*params)
-    
+
     # get the ball-park estimate
     p0 = utils.brute_force_search(response, utils.error_function, func, grids, bounds)
-    
+
     # make sure we fit right
     npt.assert_equal(params, p0[0])
-    
+
     # make sure we sliced it right
     npt.assert_equal(p0[2][0].min(),from_1)
     npt.assert_equal(p0[2][0].max(),to_1)
     npt.assert_equal(p0[2][1].min(),from_2)
     npt.assert_equal(p0[2][1].max(),to_2)
-    
-    
+
+
 
 def test_recast_estimation_results():
 
@@ -122,112 +126,112 @@ def test_recast_estimation_results():
     voxel_index = (1,2,3)
     auto_fit = True
     verbose = 1
-    
+
     # create the sweeping bar stimulus in memory
[…]
