+++ popeye/spatiotemporal.py
@@ -19,167 +19,174 @@ import popeye.utilities as utils
 from popeye.base import PopulationModel, PopulationFit
 from popeye.spinach import generate_og_receptive_field, generate_strf_timeseries, generate_rf_timeseries_nomask
 
+# Python 3 compatibility:
+try:
+    xrange
+except NameError:  # python3
+    xrange = range
+
+
 class SpatioTemporalModel(PopulationModel):
-    
+
     """
     A Gaussian population receptive field model class
-    
+
     """
-    
+
     def __init__(self, stimulus, hrf_model):
-        
+
         """
         A spatiotemporal population receptive field model.
-        
+
         Paramaters
         ----------
-        
+
         stimulus : `VisualStimulus` class object
             A class instantiation of the `VisualStimulus` class
             containing a representation of the visual stimulus.
-        
-        
+
+
         """
-        
+
         PopulationModel.__init__(self, stimulus, hrf_model)
-    
-    
+
+
     # for the final solution, we use spatiotemporal
     def generate_ballpark_prediction(self, x, y, sigma, weight, beta, baseline):
-        
+
         # generate the RF
         spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
         spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
-        
+
         # spatial_response
         rf_ts = generate_rf_timeseries_nomask(self.stimulus.stim_arr0, spatial_rf)
-        
+
         # temporal response
         m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
-        
+
         # clean up nan
         m_ts[np.isnan(m_ts)] = 0
         p_ts[np.isnan(p_ts)] = 0
-        
+
         # normalize each timeseries
         m_ts = utils.normalize(m_ts,0,1)
         p_ts = utils.normalize(p_ts,0,1)
-        
+
         # mix them
         mp_ts = (1-weight) * m_ts + weight * p_ts
-        
+
         # convolve with HRF
         hrf = self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
         model = fftconvolve(mp_ts, hrf)[0:len(mp_ts)]
-        
+
         # scale
         model *= beta
-         
+
         # offset
         model += baseline
-        
+
         return model
-    
+
     # for the final solution, we use spatiotemporal
     def generate_prediction(self, x, y, sigma, weight, beta, baseline):
-        
+
         # generate the RF
         spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x, self.stimulus.deg_y)
         spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x[0,0:2])**2)
-        
+
         # spatial response
         rf_ts = generate_rf_timeseries_nomask(self.stimulus.stim_arr, spatial_rf)
-        
+
         # temporal response
         m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
-        
+
         # clean up nan
         m_ts[np.isnan(m_ts)] = 0
         p_ts[np.isnan(p_ts)] = 0
-        
+
         # normalize each timeseries
         m_ts = utils.normalize(m_ts,0,1)
         p_ts = utils.normalize(p_ts,0,1)
-        
+
         # mix them
[…]
