### 2016-11-22 cc053b9
Update Travis to use conda.
