### 2016-11-22 91ce814
Install all dependencies upfront, streamline install section.
