+++ .travis.yml
@@ -2,7 +2,7 @@ sudo: false
 
 env:
   global:
-    - PIP_DEPS="numpy cython scipy matplotlib nibabel sharedmem statsmodels"
+    - PIP_DEPS="numpy cython scipy matplotlib nibabel sharedmem statsmodels nose python-coveralls coverage"
 
 matrix:
   include:
@@ -31,11 +31,7 @@ before_install:
 
 
 install:
-  - python setup.py install build_ext
-  - pip install python-coveralls
-  - pip install coverage       # Add me to install coverage.py
-  - pip install nose           # Add me to install nose
-  # command to run tests, e.g. python setup.py test
+  - python setup.py install
 script:
     # Change into an innocuous directory and find tests from installation
     - mkdir for_testing
