### 2016-11-22 8442575
ENH: added auditory model that uses pct sig ch rather than DC offset
