+++ popeye/auditory_bin_pct.py
@@ -1,0 +1,280 @@
+#!/usr/bin/python
+
+""" Classes and functions for fitting population encoding models """
+
+from __future__ import division
+import warnings
+warnings.simplefilter("ignore")
+
+import numpy as np
+from scipy.signal import fftconvolve
+from scipy.interpolate import interp1d
+import nibabel
+
+from popeye.onetime import auto_attr
+import popeye.utilities as utils
+from popeye.base import PopulationModel, PopulationFit
+from popeye.spinach import generate_rf_timeseries_1D, binner
+
+class AuditoryModel(PopulationModel):
+    
+    def __init__(self, stimulus, hrf_model):
+        
+        r"""A 1D Gaussian population receptive field model [1]_.
+        
+        Paramaters
+        ----------
+        
+        stimulus : `AuditoryStimulus` class object
+            A class instantiation of the `AuditoryStimulus` class
+            containing a representation of the auditory stimulus.
+            
+        hrf_model : callable
+            A function that generates an HRF model given an HRF delay.
+            For more information, see `popeye.utilties.double_gamma_hrf_hrf`
+        
+        
+        References
+        ----------
+        
+        .. [1] Thomas JM, Huber E, Stecker GC, Boynton GM, Saenz M, Fine I. (2015) 
+        Population receptive field estimates in human auditory cortex. 
+        NeuroImage 105, 428-439.
+        
+        """
+        
+        # invoke the base class
+        PopulationModel.__init__(self, stimulus, hrf_model)
+    
+    def generate_prediction(self, center_freq, sigma, beta, hrf_delay):
+        
+        r"""
+        Generate a prediction for the 1D Gaussian model.
+        
+        This function generates a prediction of the 1D Gaussian model, 
+        given a stimulus and the stimulus-referred model parameters.  This
+        function operates on the native stimulus.
+        
+        Paramaters
+        ----------
+        
+        center_freq : float
+            The center frequency of the 1D Gaussian, units are in Hz.
+            
+        sigma : float
+            The dispersion of the 1D Gaussian, units are in Hz.
+        
+        beta : float
+            The scaling factor to account for arbitrary units of the BOLD signal.
+        
+        hrf_delay : float
+            The delay of the HRF, units are in seconds.
+        
+        """
+        
+        # generate stimulus time-series
+        rf = self.receptive_field(center_freq, sigma)
+        
+        # if the tuning curve is running off the edges of the frequency space
+        if self.compare_rf_volumes(rf):
+            return np.inf
+        
+        # create mask for speed
+        distance = self.stimulus.freqs - center_freq
+        mask = np.zeros_like(distance, dtype='uint8')
+        mask[distance < (5*sigma)] = 1
+        
+        # extract the response
+        response = generate_rf_timeseries_1D(self.stimulus.spectrogram, rf, mask)
+        
+        # bin the response to the resolution of the BOLD data
+        binned_response = self.binned_response(response)
+        
+        # roll it forward to correct for the binning
+        rolled_response = np.roll(binned_response, 0)
+        
+        # generate the HRF
+        hrf = utils.double_gamma_hrf(hrf_delay, self.stimulus.tr_length)
+        
+        # pad and convolve
+        model = fftconvolve(rolled_response, hrf)[0:len(rolled_response)]
+        
+        # convert units
+        model = (model - np.mean(model))/np.mean(model)
+        
+        # scale by beta
+        model *= beta
+        
+        return model
+    
+    def generate_ballpark_prediction(self, center_freq, sigma, beta, hrf_delay):
+        
+        r"""
+        Generate a prediction for the 1D Gaussian model.
+        
+        This function generates a prediction of the 1D Gaussian model, 
[…]
