### 2016-11-22 4285a72
TST: added test for new auditory model using pct sig ch rather than DC offset
