### 2016-11-22 59b26cd
RM: removed the info, upstaged by version
