### 2016-11-22 ef1ea7e
ENH: added dependencies to the setup for pypi
