+++ setup.py
@@ -45,7 +45,7 @@ opts = dict(name=NAME,
             version=VERSION,
             packages=PACKAGES,
             package_data=PACKAGE_DATA,
-            requires=REQUIRES,
+            requires=['scipy','numpy','matplotlib','statsmodels','nibabel','sharedmem','cython'],
             )
 
 try:
