### 2016-11-22 217e551
Use pytest instead of nose for testing.
