### 2016-11-22 085a1a0
Use pytest on circle as well.
