+++ .travis.yml
@@ -29,7 +29,6 @@ before_install:
   - conda install python=$PYTHON_VERSION
   - pip install $PIP_DEPS
 
-
 install:
   - python setup.py install
 script:
