### 2016-11-23 4d78e35
ENH: added visual field masking to the model baseclass. if no model.mask_size is given, then the entire visual field is evaluated. otherwise, we evaluate the model at mask_size*sigma
