### 2016-11-23 b22ca74
RM: removed the masked/nonmasked variations since each model is now agnnostic as to masking and since users will determine that on a case-by-case basis at runtime
