+++ popeye/spatiotemporal_css_masked.py
@@ -1,305 +1,0 @@
-#!/usr/bin/python
-
-""" Classes and functions for fitting SpatioTemporal population receptive field models """
-
-from __future__ import division, print_function, absolute_import
-import time
-import gc
-import warnings
-warnings.simplefilter("ignore")
-
-import numpy as np
-from scipy.stats import linregress
-from scipy.signal import fftconvolve
-from scipy.integrate import trapz, simps
-import nibabel
-
-from popeye.onetime import auto_attr
-import popeye.utilities as utils
-from popeye.base import PopulationModel, PopulationFit
-from popeye.spinach import generate_og_receptive_field, generate_strf_timeseries, generate_rf_timeseries
-
-class SpatioTemporalModel(PopulationModel):
-    
-    """
-    A Gaussian population receptive field model class
-    
-    """
-    
-    def __init__(self, stimulus, hrf_model):
-        
-        """
-        A spatiotemporal population receptive field model.
-        
-        Paramaters
-        ----------
-        
-        stimulus : `VisualStimulus` class object
-            A class instantiation of the `VisualStimulus` class
-            containing a representation of the visual stimulus.
-        
-        
-        """
-        
-        PopulationModel.__init__(self, stimulus, hrf_model)
-    
-    
-    # for the final solution, we use spatiotemporal
-    def generate_ballpark_prediction(self, x, y, sigma, n, weight, beta, baseline):
-        
-        # mask for speed
-        mask = self.distance_mask_coarse(x, y, sigma*6)
-        
-        # generate the RF
-        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
-        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
-        
-        # spatial_response
-        rf_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
-        
-        # compression
-        rf_ts **= n
-        
-        # temporal response
-        m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
-        
-        # clean up nan
-        m_ts[np.isnan(m_ts)] = 0
-        p_ts[np.isnan(p_ts)] = 0
-        
-        # normalize each timeseries
-        m_ts = utils.normalize(m_ts,0,1)
-        p_ts = utils.normalize(p_ts,0,1)
-        
-        # mix them
-        mp_ts = (1-weight) * m_ts + weight * p_ts
-        
-        # convolve with HRF
-        hrf = self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
-        model = fftconvolve(mp_ts, hrf)[0:len(mp_ts)]
-        
-        # scale
-        model *= beta
-         
-        # offset
-        model += baseline
-        
-        return model
-    
-    # for the final solution, we use spatiotemporal
-    def generate_prediction(self, x, y, sigma, n, weight, beta, baseline):
-        
-        # mask for speed
-        mask = self.distance_mask(x, y, sigma*6)
-        
-        # generate the RF
-        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x, self.stimulus.deg_y)
-        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x[0,0:2])**2)
-        
-        # spatial response
[…]
