### 2016-11-23 3926b31
RF: added pct sig ch units and now abiding by the new visual masking from model base class
