### 2016-11-23 47bad64
TST: updated test for new base class
