### 2016-11-23 e732d92
TST: updated tests for new base api
