### 2016-11-23 6acb08d
TST: updated tests for new base api
