### 2016-11-23 fc75e4e
RF: udpated models to reflect new base api
