+++ popeye/ogb.py
@@ -103,9 +103,9 @@ class GaussianFit(PopulationFit):
     
     """
     
-    def __init__(self, model, data, grids, bounds, Ns,
-                 voxel_index=(1,2,3), auto_fit=True, verbose=0):
-        
+    def __init__(self, model, data, grids, bounds,
+                 voxel_index=(1,2,3), Ns=None, auto_fit=True, verbose=0):
+                 
         r""" A class containing tools for fitting the 2D Gaussian pRF model.
         
         The `GaussianFit` class houses all the fitting tool that are associated with 
@@ -163,8 +163,8 @@ class GaussianFit(PopulationFit):
         
         """
         
-        PopulationFit.__init__(self, model, data, grids, bounds, Ns, 
-                               voxel_index, auto_fit, verbose)
+        PopulationFit.__init__(self, model, data, grids, bounds, 
+                               voxel_index, Ns, auto_fit, verbose)
 
     
     @auto_attr
