### 2016-11-23 bb8bbf0
RF: beautifying spacing etc
