+++ popeye/spatiotemporal_css.py
@@ -17,7 +17,7 @@ import nibabel
 from popeye.onetime import auto_attr
 import popeye.utilities as utils
 from popeye.base import PopulationModel, PopulationFit
-from popeye.spinach import generate_og_receptive_field, generate_strf_timeseries, generate_rf_timeseries_nomask
+from popeye.spinach import generate_og_receptive_field, generate_mp_timeseries, generate_rf_timeseries
 
 # Python 3 compatibility:
 try:
@@ -27,265 +27,242 @@ except NameError:  # python3
 
 
 class SpatioTemporalModel(PopulationModel):
-
+    
     """
     A Gaussian population receptive field model class
-
+    
     """
-
+    
     def __init__(self, stimulus, hrf_model):
-
+        
         """
         A spatiotemporal population receptive field model.
-
+        
         Paramaters
         ----------
-
+        
         stimulus : `VisualStimulus` class object
             A class instantiation of the `VisualStimulus` class
             containing a representation of the visual stimulus.
-
-
+            
+            
         """
-
+        
         PopulationModel.__init__(self, stimulus, hrf_model)
-
-
+        
+        
     # for the final solution, we use spatiotemporal
-    def generate_ballpark_prediction(self, x, y, sigma, n, weight, beta, baseline):
-
+    def generate_ballpark_prediction(self, x, y, sigma, n, weight, beta):
+        
+        # mask for speed
+        mask = self.distance_mask_coarse(x, y, sigma)
+        
         # generate the RF
         spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
         spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
-
+        
         # spatial_response
-        rf_ts = generate_rf_timeseries_nomask(self.stimulus.stim_arr0, spatial_rf)
-
+        spatial_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
+        
         # compression
-        rf_ts **= n
-
+        spatial_ts **= n
+        
         # temporal response
-        m_ts,p_ts = generate_strf_timeseries(rf_ts,self.m_resp,self.p_resp,self.stimulus.flicker_vec)
-
-        # clean up nan
-        m_ts[np.isnan(m_ts)] = 0
-        p_ts[np.isnan(p_ts)] = 0
-
-        # normalize each timeseries
-        m_ts = utils.normalize(m_ts,0,1)
-        p_ts = utils.normalize(p_ts,0,1)
-
+        m_ts, p_ts = generate_mp_timeseries(spatial_ts, self.m_amp, self.p_amp, self.stimulus.flicker_vec)
+        
         # mix them
         mp_ts = (1-weight) * m_ts + weight * p_ts
-
+        
         # convolve with HRF
         hrf = self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
         model = fftconvolve(mp_ts, hrf)[0:len(mp_ts)]
-
+        
+        # convert units
+        model = (model - np.mean(model)) / np.mean(model)
+        
         # scale
         model *= beta
-
-        # offset
-        model += baseline
-
+        
         return model
-
+        
     # for the final solution, we use spatiotemporal
-    def generate_prediction(self, x, y, sigma, n, weight, beta, baseline):
-
+    def generate_prediction(self, x, y, sigma, n, weight, beta):
+        
+        # mask for speed
[…]
