### 2016-11-23 cd93d62
RF: complying with new base api
