+++ popeye/spatiotemporal_betas.py
@@ -1,0 +1,271 @@
+#!/usr/bin/python
+
+""" Classes and functions for fitting SpatioTemporal population receptive field models """
+
+from __future__ import division, print_function, absolute_import
+import time
+import gc
+import warnings
+warnings.simplefilter("ignore")
+
+import numpy as np
+from scipy.stats import linregress
+from scipy.signal import fftconvolve
+from scipy.integrate import trapz, simps
+import nibabel
+
+from popeye.onetime import auto_attr
+import popeye.utilities as utils
+from popeye.base import PopulationModel, PopulationFit
+from popeye.spinach import generate_og_receptive_field, generate_mp_timeseries, generate_rf_timeseries
+
+class SpatioTemporalModel(PopulationModel):
+    
+    """
+    A Gaussian population receptive field model class
+    
+    """
+    
+    def __init__(self, stimulus, hrf_model):
+        
+        """
+        A spatiotemporal population receptive field model.
+        
+        Paramaters
+        ----------
+        
+        stimulus : `VisualStimulus` class object
+            A class instantiation of the `VisualStimulus` class
+            containing a representation of the visual stimulus.
+        
+        
+        """
+        
+        PopulationModel.__init__(self, stimulus, hrf_model)
+    
+    def generate_mp_timeseries_long(self, spatial_ts):
+        
+        m_ts = np.zeros(self.stimulus.flicker_vec.shape[0])
+        m_ts[self.stimulus.flicker_vec==1] = self.m_amp[0] * spatial_ts[self.stimulus.flicker_vec==1]
+        m_ts[self.stimulus.flicker_vec==2] = self.m_amp[1] * spatial_ts[self.stimulus.flicker_vec==2]
+        
+        p_ts = np.zeros(self.stimulus.flicker_vec.shape[0])
+        p_ts[self.stimulus.flicker_vec==1] = self.p_amp[0] * spatial_ts[self.stimulus.flicker_vec==1]
+        p_ts[self.stimulus.flicker_vec==2] = self.p_amp[1] * spatial_ts[self.stimulus.flicker_vec==2]
+        
+        return m_ts, p_ts
+        
+    
+    # for the final solution, we use spatiotemporal
+    def generate_ballpark_prediction(self, x, y, sigma, mbeta, pbeta):
+        
+        # mask for speed
+        mask = self.distance_mask_coarse(x, y, sigma)
+        
+        # generate the RF
+        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
+        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
+        
+        # spatial_response
+        spatial_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
+        
+        # temporal response
+        m_ts, p_ts = generate_mp_timeseries(spatial_ts, self.m_amp, self.p_amp, self.stimulus.flicker_vec)
+        
+        # convolve with HRF
+        hrf = self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
+        
+        # M
+        m_model = fftconvolve(m_ts, hrf)[0:len(m_ts)]
+        
+        # P
+        p_model = fftconvolve(p_ts, hrf)[0:len(p_ts)]
+        
+        # convert units
+        m_model = (m_model - np.mean(m_model))/np.mean(m_model)
+        p_model = (p_model - np.mean(p_model))/np.mean(p_model)
+        
+        # mix
+        model = m_model * mbeta + p_model * pbeta
[…]
