### 2016-11-23 1e97043
TST: removing test coverage for conditionals of noninterest
