### 2016-11-23 3274b0a
TST: removing test coverage for conditionals of noninterest
