+++ README.md
@@ -7,9 +7,13 @@
 popeye
 ======
 
+![popeye](/images/lgn.png)
+
 popeye is a Python module for estimating population receptive fields
 from from fMRI data built on top of SciPy and distributed under the
 3-Clause BSD license.
+
+Documentation for popeye and contact information for the authors are available at https://kdesimone.github.io/popeye
 
 popeye is currently under development.
 
