### 2016-11-23 9516775
RF: fixed path to image
