### 2016-11-23 8b456e9
DOC: removed figs
