+++ README.md
@@ -7,8 +7,6 @@
 popeye
 ======
 
-![popeye](/docs/images/lgn.png)
-
 popeye is a Python module for estimating population receptive fields
 from from fMRI data built on top of SciPy and distributed under the
 3-Clause BSD license.
