### 2016-11-23 8896d9a
DOC: added pypi install
