### 2016-11-25 689732e
DOC: removed .dev from version_extra
