+++ popeye/version.py
@@ -5,7 +5,7 @@
 _version_major = 0
 _version_minor = 1
 _version_micro = 0
-_version_extra = '.dev'
+# _version_extra = '.dev'
 #_version_extra = ''
 
 # Construct full version string from these.
