### 2016-11-26 20ac3a0
Replace all usage of nose.tools with numpy.testing. assert_true => assert_
