+++ popeye/tests/test_ogb.py
@@ -4,7 +4,6 @@ import ctypes
 
 import numpy as np
 import numpy.testing as npt
-import nose.tools as nt
 from scipy.signal import fftconvolve
 
 import popeye.utilities as utils
@@ -12,7 +11,7 @@ import popeye.ogb as ogb
 from popeye.visual_stimulus import VisualStimulus, simulate_bar_stimulus
 from popeye.spinach import generate_og_receptive_field
 def test_ogb_fit():
-    
+
     # stimulus features
     viewing_distance = 38
     screen_width = 25
@@ -30,18 +29,18 @@ def test_ogb_fit():
     voxel_index = (1,2,3)
     auto_fit = True
     verbose = 1
-    
+
     # create the sweeping bar stimulus in memory
-    bar = simulate_bar_stimulus(pixels_across, pixels_down, viewing_distance, 
+    bar = simulate_bar_stimulus(pixels_across, pixels_down, viewing_distance,
                                 screen_width, thetas, num_bar_steps, num_blank_steps, ecc)
-    
+
     # create an instance of the Stimulus class
     stimulus = VisualStimulus(bar, viewing_distance, screen_width, scale_factor, tr_length, dtype)
-    
+
     # initialize the gaussian model
     model = ogb.GaussianModel(stimulus, utils.double_gamma_hrf)
     model.tr_length = 1.0
-    
+
     # generate a random pRF estimate
     x = -2.2
     y = 2.6
@@ -49,10 +48,10 @@ def test_ogb_fit():
     beta = 85
     baseline = 0.5
     hrf_delay = -0.3
-    
+
     # create data
     data = model.generate_prediction(x, y, sigma, beta, baseline, hrf_delay)
-    
+
     # set search grid
     x_grid = (-5,5)
     y_grid = (-5,5)
@@ -60,7 +59,7 @@ def test_ogb_fit():
     b_grid = (1,100)
     bl_grid = (-1,1)
     h_grid = (-4.0,4.0)
-    
+
     # set search bounds
     x_bound = (-12.0,12.0)
     y_bound = (-12.0,12.0)
@@ -68,33 +67,32 @@ def test_ogb_fit():
     b_bound = (1e-8,None)
     bl_bound = (-1,1)
     h_bound = (-5.0,5.0)
-    
+
     # loop over each voxel and set up a GaussianFit object
     grids = (x_grid, y_grid, s_grid, b_grid, bl_grid, h_grid)
     bounds = (x_bound, y_bound, s_bound, b_bound, bl_bound, h_bound)
-    
+
     # fit the response
     fit = ogb.GaussianFit(model, data, grids, bounds, Ns=Ns)
-    
+
     # test ballpark
-    nt.assert_almost_equal(fit.x0, 0, 1)
-    nt.assert_almost_equal(fit.y0, 0, 1)
-    nt.assert_almost_equal(fit.s0, 2.8, 1)
-    nt.assert_almost_equal(fit.beta0, 100, 1)
-    nt.assert_almost_equal(fit.baseline0, -1, 1)
-    nt.assert_almost_equal(fit.hrf0, 0, 1)
-    
+    npt.assert_almost_equal(fit.x0, 0, 1)
+    npt.assert_almost_equal(fit.y0, 0, 1)
+    npt.assert_almost_equal(fit.s0, 2.8, 1)
+    npt.assert_almost_equal(fit.beta0, 100, 1)
+    npt.assert_almost_equal(fit.baseline0, -1, 1)
+    npt.assert_almost_equal(fit.hrf0, 0, 1)
+
     # test final estimate
-    nt.assert_almost_equal(fit.x, x, 1)
[…]
