### 2016-11-26 494f824
RM: tossing out defunct model
