### 2016-11-26 31fefeb
RF: abiding by new masking in model api
