+++ popeye/spatiotemporal_css_betas.py
@@ -1,0 +1,277 @@
+#!/usr/bin/python
+
+""" Classes and functions for fitting SpatioTemporal population receptive field models """
+
+from __future__ import division, print_function, absolute_import
+import time
+import gc
+import warnings
+warnings.simplefilter("ignore")
+
+import numpy as np
+from scipy.stats import linregress
+from scipy.signal import fftconvolve
+from scipy.integrate import trapz, simps
+import nibabel
+
+from popeye.onetime import auto_attr
+import popeye.utilities as utils
+from popeye.base import PopulationModel, PopulationFit
+from popeye.spinach import generate_og_receptive_field, generate_mp_timeseries, generate_rf_timeseries
+
+class SpatioTemporalModel(PopulationModel):
+    
+    """
+    A Gaussian population receptive field model class
+    
+    """
+    
+    def __init__(self, stimulus, hrf_model):
+        
+        """
+        A spatiotemporal population receptive field model.
+        
+        Paramaters
+        ----------
+        
+        stimulus : `VisualStimulus` class object
+            A class instantiation of the `VisualStimulus` class
+            containing a representation of the visual stimulus.
+        
+        
+        """
+        
+        PopulationModel.__init__(self, stimulus, hrf_model)
+    
+    
+    # for the final solution, we use spatiotemporal
+    def generate_ballpark_prediction(self, x, y, sigma, n, mbeta, pbeta):
+        
+        # mask for speed
+        mask = self.distance_mask_coarse(x, y, sigma)
+        
+        # generate the RF
+        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
+        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
+        
+        # spatial_response
+        spatial_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
+        
+        # compression
+        spatial_ts **= n
+        
+        # temporal response
+        m_ts, p_ts = generate_mp_timeseries(spatial_ts, self.m_amp, self.p_amp, self.stimulus.flicker_vec)
+        
+        # convolve with HRF
+        hrf = self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
+        
+        # M
+        m_model = fftconvolve(m_ts, hrf)[0:len(m_ts)]
+        
+        # P
+        p_model = fftconvolve(p_ts, hrf)[0:len(p_ts)]
+        
+        # convert units
+        m_model = (m_model - np.mean(m_model))/np.mean(m_model)
+        p_model = (p_model - np.mean(p_model))/np.mean(p_model)
+        
+        # mix
+        model = m_model * mbeta + p_model * pbeta
+        
+        return model
+    
+    # for the final solution, we use spatiotemporal
+    def generate_prediction(self, x, y, sigma, n, mbeta, pbeta):
+        
+        # mask for speed
+        mask = self.distance_mask_coarse(x, y, sigma)
+        
+        # generate the RF
+        spatial_rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
+        spatial_rf /= ((2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2)
+        
+        # spatial_response
+        spatial_ts = generate_rf_timeseries(self.stimulus.stim_arr0, spatial_rf, mask)
+        
+        # compression
[…]
