### 2016-11-27 #103 @cMadan
@tknapen - thanks for your help reviewing this, it looks like nearly everything is in order for the paper to be accepted.

@kdesimone - as noted, the version number doesn't match with what was submitted to JOSS. Would it be fair to refer to the current code as v.0.1.0, or do you mean v.0.1.0.dev to be distinct from that?

After that is resolved, I think we just need the code to be archived in a repository such as figshare or zenodo and we'd be all set. Am I missing anything, @arfon?
