### 2016-11-27 #103 @arfon
> After that is resolved, I think we just need the code to be archived in a repository such as figshare or zenodo and we'd be all set. Am I missing anything, @arfon?

👍 that's correct @cMadan. Once the review is complete @kdesimone please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive.
