### 2016-11-28 75e8463
ENH: added ROI badge
