### 2016-11-28 cfad646
RF: updated the version
