+++ popeye/version.py
@@ -3,7 +3,7 @@
 # Format expected by setup.py and doc/source/conf.py: string of form "X.Y.Z"
 
 _version_major = 0
-_version_minor = 1
+_version_minor = 2
 _version_micro = 0
 # _version_extra = '.dev'
 _version_extra = ''
