### 2016-11-28 #103 @kdesimone
I actually incremented the version number to 0.2.0 when I added the DOI tag.
