+++ popeye/css_nohrf.py
@@ -50,7 +50,7 @@ class CompressiveSpatialSummationModel(PopulationModel):
         
     
     # main method for deriving model time-series
-    def generate_ballpark_prediction(self, x, y, sigma, n, beta, baseline):
+    def generate_ballpark_prediction(self, x, y, sigma, n):
         
         # generate the RF
         rf = generate_og_receptive_field(x, y, sigma,self.stimulus.deg_x0, self.stimulus.deg_y0)
@@ -70,16 +70,16 @@ class CompressiveSpatialSummationModel(PopulationModel):
         # convolve it with the stimulus
         model = fftconvolve(response, hrf)[0:len(response)]
         
+        # convert units
+        model = (model = np.mean(model)) / np.mean(model)
+        
         # scale it by beta
         model *= beta
-        
-        # offset
-        model += baseline
         
         return model
         
     # main method for deriving model time-series
-    def generate_prediction(self, x, y, sigma, n, beta, baseline):
+    def generate_prediction(self, x, y, sigma, n, beta):
         
         # generate the RF
         rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x, self.stimulus.deg_y)
@@ -99,11 +99,11 @@ class CompressiveSpatialSummationModel(PopulationModel):
         # convolve it with the stimulus
         model = fftconvolve(response, hrf)[0:len(response)]
         
+        # convert units
+        model = (model = np.mean(model)) / np.mean(model)
+        
         # scale it by beta
         model *= beta
-        
-        # offset
-        model += baseline
         
         return model
         
@@ -114,10 +114,9 @@ class CompressiveSpatialSummationFit(PopulationFit):
     
     """
     
-    def __init__(self, model, data, grids, bounds, 
+    def __init__(self, model, data, grids, bounds,
                  voxel_index=(1,2,3), Ns=None, auto_fit=True, verbose=0):
         
-
         
         r"""
         A class containing tools for fitting the CSS pRF model.
@@ -204,11 +203,7 @@ class CompressiveSpatialSummationFit(PopulationFit):
     @auto_attr
     def beta0(self):
         return self.ballpark[4]
-    
-    @auto_attr
-    def baseline0(self):
-        return self.ballpark[5]
-    
+        
     @auto_attr
     def x(self):
         return self.estimate[0]
@@ -228,10 +223,6 @@ class CompressiveSpatialSummationFit(PopulationFit):
     @auto_attr
     def beta(self):
         return self.estimate[4]
-        
-    @auto_attr
-    def baseline(self):
-        return self.estimate[5]
         
     @auto_attr
     def rho(self):
