### 2016-11-28 0a21336
TST: updated tests
