### 2016-11-28 a0d3d27
TST: updated tests
