+++ popeye/tests/test_css_nohrf.py
@@ -4,7 +4,6 @@ import ctypes
 
 import numpy as np
 import numpy.testing as npt
-import nose.tools as nt
 
 import popeye.utilities as utils
 import popeye.css_nohrf as css
@@ -46,10 +45,9 @@ def test_css_fit():
     sigma = 1.23
     n = 0.90
     beta = 1.0
-    baseline = 0.0
     
     # create the "data"
-    data =  model.generate_prediction(x, y, sigma, n, beta, baseline)
+    data =  model.generate_prediction(x, y, sigma, n, beta)
     
     # set search grid
     x_grid = (-3,2)
@@ -57,7 +55,6 @@ def test_css_fit():
     s_grid = (1/stimulus.ppd,2.75)
     n_grid = (0.1,0.90)
     b_grid = (0.01,1)
-    bl_grid = (-2,2)
     h_grid = (-4.0,4.0)
     
     # set search bounds
@@ -66,31 +63,28 @@ def test_css_fit():
     s_bound = (1/stimulus.ppd,10)
     n_bound = (1e-8,1.0)
     b_bound = (1e-8,1e5)
-    bl_bound = (None,None)
     h_bound = (-5.0,5.0)
     
     # loop over each voxel and set up a GaussianFit object
-    grids = (x_grid, y_grid, s_grid, n_grid, b_grid, bl_grid,)
-    bounds = (x_bound, y_bound, s_bound, n_bound, b_bound, bl_bound,)
+    grids = (x_grid, y_grid, s_grid, n_grid, b_grid,)
+    bounds = (x_bound, y_bound, s_bound, n_bound, b_bound,)
     
     # fit the response
     fit = css.CompressiveSpatialSummationFit(model, data, grids, bounds, Ns=Ns)
     
     # fine
-    nt.assert_almost_equal(fit.x, x, 1)
-    nt.assert_almost_equal(fit.y, y, 1)
-    nt.assert_almost_equal(fit.sigma, sigma, 1)
-    nt.assert_almost_equal(fit.n, n, 1)
-    nt.assert_almost_equal(fit.beta, beta, 1)
-    nt.assert_almost_equal(fit.baseline, baseline, 1)
+    npt.assert_almost_equal(fit.x, x, 1)
+    npt.assert_almost_equal(fit.y, y, 1)
+    npt.assert_almost_equal(fit.sigma, sigma, 1)
+    npt.assert_almost_equal(fit.n, n, 1)
+    npt.assert_almost_equal(fit.beta, beta, 1)
     
     # coarse
-    npt.assert_almost_equal((fit.x0,fit.y0,fit.s0,fit.n0,fit.beta0,fit.baseline0),(2.0, -0.5, 0.72833937882323319, 0.10000000000000001, 1.0, 0.0))
+    npt.assert_almost_equal(fit.ballpark,[-0.5       , -0.5       ,  0.72833938,  0.9       ,  0.505     ])
     
     # overloaded
-    npt.assert_almost_equal(fit.overloaded_estimate,[2.5272803327899305,
-                                                     2.7411676344232436,
-                                                     1.2965338406630695,
-                                                     0.89999999997351177,
-                                                     0.99999999998062961,
-                                                     3.5926082629195817e-13])
+    npt.assert_almost_equal(fit.overloaded_estimate,[2.5272803327911109,
+                                                     2.741167634420512,
+                                                     1.2965338406813691,
+                                                     0.90000000004104885,
+                                                     1.0000000000009019])
