### 2016-11-28 0664809
ENH: added hrf generation to the PopulationModel base class
