+++ popeye/base.py
@@ -99,6 +99,12 @@ class PopulationModel(object):
             
         return mask
     
+    def hrf(self):
+        if hasattr(self, 'hrf_delay'): # pragma: no cover
+            return self.hrf_model(self.hrf_delay, self.stimulus.tr_length)
+        else:
+            raise NotImplementedError("You must set the HRF delay to generate the HRF")
+    
 class PopulationFit(object):
     
     
