### 2016-11-28 d4ae439
RF: updating spatiotemporal pRF model to use new hrf generation from base class
