### 2016-11-28 4683561
RF: normalizing the spm_hrf by the integral so we offset area under curver for variable hrf_delays
