### 2016-11-29 b705d3d
TST: exstending coverage to include parameters in Fit object
