### 2016-11-29 #103 @cMadan
@kdesimone - good to know, thanks!

@arfon - how should I change the version number? (I assume not just editing the first post in this issue thread?) Apart from that, we're all set here!
