### 2016-11-29 d8cf6a1
RF: moved HRF fitting out of og, that is for another model
