### 2016-11-29 92e3594
ENH: new og model using voxelwise HRF fitting
