+++ popeye/og_hrf.py
@@ -1,0 +1,223 @@
+#!/usr/bin/python
+
+""" Classes and functions for estimating Gaussian pRF model """
+
+from __future__ import division
+import warnings
+warnings.simplefilter("ignore")
+
+import numpy as np
+from scipy.signal import fftconvolve
+import nibabel
+
+from popeye.onetime import auto_attr
+import popeye.utilities as utils
+from popeye.base import PopulationModel, PopulationFit
+from popeye.spinach import generate_og_receptive_field, generate_rf_timeseries
+
+class GaussianModel(PopulationModel):
+    
+    def __init__(self, stimulus, hrf_model, nuisance=None):
+        
+        r"""A 2D Gaussian population receptive field model [1]_.
+        
+        Paramaters
+        ----------
+        
+        stimulus : `VisualStimulus` class object
+            A class instantiation of the `VisualStimulus` class
+            containing a representation of the visual stimulus.
+            
+        hrf_model : callable
+            A function that generates an HRF model given an HRF delay.
+            For more information, see `popeye.utilties.double_gamma_hrf_hrf`
+        
+        
+        References
+        ----------
+        
+        .. [1] Dumoulin SO, Wandell BA. (2008) Population receptive field 
+        estimates in human visual cortex. NeuroImage 39:647-660
+        
+        """
+        
+        PopulationModel.__init__(self, stimulus, hrf_model, nuisance)
+    
+    # main method for deriving model time-series
+    def generate_ballpark_prediction(self, x, y, sigma, beta, hrf_delay):
+        
+        # mask for speed
+        mask = self.distance_mask_coarse(x, y, sigma)
+        
+        # generate the RF
+        rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x0, self.stimulus.deg_y0)
+        rf /= (2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x0[0,0:2])**2
+                
+        # extract the stimulus time-series
+        response = generate_rf_timeseries(self.stimulus.stim_arr0, rf, mask)
+        
+        # convolve it with the stimulus
+        model = fftconvolve(response, self.hrf_model(hrf_delay, self.stimulus.tr_length))[0:len(response)]
+        
+        # convert units
+        model = (model-np.mean(model)) / np.mean(model)
+        
+        # scale it by beta
+        model *= beta
+        
+        return model
+        
+    # main method for deriving model time-series
+    def generate_prediction(self, x, y, sigma, beta, hrf_delay):
+        
+        # mask for speed
+        mask = self.distance_mask(x, y, sigma)
+        
+        # generate the RF
+        rf = generate_og_receptive_field(x, y, sigma, self.stimulus.deg_x, self.stimulus.deg_y)
+        rf /= (2 * np.pi * sigma**2) * 1/np.diff(self.stimulus.deg_x[0,0:2])**2
+        
+        # extract the stimulus time-series
+        response = generate_rf_timeseries(self.stimulus.stim_arr, rf, mask)
+        
+        # convolve it with the stimulus
+        model = fftconvolve(response, self.hrf_model(hrf_delay, self.stimulus.tr_length))[0:len(response)]
+        
+        # convert units
+        model = (model-np.mean(model)) / np.mean(model)
+        
+        # scale it by beta
+        model *= beta
+        
+        return model
+        
+class GaussianFit(PopulationFit):
+    
+    r"""
+    A 2D Gaussian population receptive field fit class.
+    
+    """
+    
+    def __init__(self, model, data, grids, bounds,
+                 voxel_index=(1,2,3), Ns=None, auto_fit=True, verbose=0):
+        
+        r"""
[…]
