### 2016-11-29 fc463c8
ENH: added spatiotemporal model with voxelwise HRF fitting
