### 2016-11-29 00c713b
RF: made spatiotemporal model not correct m/p amp.  its au naturel
