+++ popeye/spatiotemporal.py
@@ -65,7 +65,7 @@ class SpatioTemporalModel(PopulationModel):
         mp_ts = (1-weight) * m_ts + weight * p_ts
         
         # convolve with HRF
-        model = fftconvolve(mp_ts, self.hrf)[0:len(mp_ts)]
+        model = fftconvolve(mp_ts, self.hrf())[0:len(mp_ts)]
         
         # convert units
         model = (model - np.mean(model)) / np.mean(model)
@@ -95,7 +95,7 @@ class SpatioTemporalModel(PopulationModel):
         mp_ts = (1-weight) * m_ts + weight * p_ts 
         
         # convolve with HRF
-        model = fftconvolve(mp_ts, self.hrf)[0:len(mp_ts)]
+        model = fftconvolve(mp_ts, self.hrf())[0:len(mp_ts)]
         
         # convert units
         model = (model - np.mean(model)) / np.mean(model)
@@ -143,13 +143,18 @@ class SpatioTemporalModel(PopulationModel):
     @auto_attr
     def m_resp(self):
         m_resp = fftconvolve(self.flickers,self.m[:,np.newaxis])
-        m_resp = utils.normalize(m_resp,-1,1)
+        # m_resp[:,0] = utils.normalize(m_resp[:,0],-1,1)
+        # m_resp[:,1] = utils.normalize(m_resp[:,1],-1,1)
+        m_resp= utils.normalize(m_resp,-1,1)
+        
         return m_resp
         
     def generate_m_resp(self, tau):
         m_rf = self.m_rf(tau)
         m_resp = fftconvolve(self.flickers,m_rf[:,np.newaxis])
-        m_resp = utils.normalize(m_resp,-1,1)
+        # m_resp[:,0] = utils.normalize(m_resp[:,0],-1,1)
+        # m_resp[:,1] = utils.normalize(m_resp[:,1],-1,1)
+        m_resp= utils.normalize(m_resp,-1,1)
         return m_resp
         
     @auto_attr
