### 2016-11-29 4812ed8
RF: made spatiotemporal_css model not correct m/p amp.  its au naturel
