### 2016-11-29 ebfd8e0
TST: updated tests for new OG new HRF
