+++ popeye/tests/test_utilities.py
@@ -126,82 +126,77 @@ def test_recast_estimation_results():
     voxel_index = (1,2,3)
     auto_fit = True
     verbose = 1
-
+    
     # create the sweeping bar stimulus in memory
     bar = simulate_bar_stimulus(pixels_across, pixels_down, viewing_distance,
                                 screen_width, thetas, num_bar_steps, num_blank_steps, ecc)
-
+                                
     # create an instance of the Stimulus class
     stimulus = VisualStimulus(bar, viewing_distance, screen_width, scale_factor, tr_length, dtype)
-
+    
     # initialize the gaussian model
-    model = og.GaussianModel(stimulus, utils.double_gamma_hrf)
-
+    model = og.GaussianModel(stimulus, utils.spm_hrf)
+    model.hrf_delay = 0
+    
     # generate a random pRF estimate
     x = -5.24
     y = 2.58
     sigma = 1.24
     beta = 2.5
-    hrf_delay = -0.25
-
+    
     # create the "data"
-    data = model.generate_prediction(x, y, sigma, beta, hrf_delay)
-
+    data = model.generate_prediction(x, y, sigma, beta)
+    
     # set search grid
     x_grid = utils.grid_slice(-5,4,5)
     y_grid = utils.grid_slice(-5,7,5)
     s_grid = utils.grid_slice(1/stimulus.ppd,5.25,5)
     b_grid = utils.grid_slice(0.1,4.0,5)
-    h_grid = utils.grid_slice(-4.0,4.0,5)
-
+    
     # set search bounds
     x_bound = (-12.0,12.0)
     y_bound = (-12.0,12.0)
     s_bound = (1/stimulus.ppd,12.0)
     b_bound = (1e-8,1e2)
-    h_bound = (-5.0,5.0)
-
+    
     # loop over each voxel and set up a GaussianFit object
-    grids = (x_grid, y_grid, s_grid, b_grid, h_grid)
-    bounds = (x_bound, y_bound, s_bound, b_bound, h_bound)
-
+    grids = (x_grid, y_grid, s_grid, b_grid,)
+    bounds = (x_bound, y_bound, s_bound, b_bound,)
+    
     # create 3 voxels of data
     all_data = np.array([data,data,data])
     indices = [(0,0,0),(0,0,1),(0,0,2)]
-
+    
     # bundle the voxels
     bundle = utils.multiprocess_bundle(og.GaussianFit, model, all_data, grids, bounds, indices)
-
+    
     # run analysis
     with sharedmem.Pool(np=3) as pool:
         output = pool.map(utils.parallel_fit, bundle)
-
+        
     # create grid parent
     arr = np.zeros((1,1,3))
     grid_parent = nibabel.Nifti1Image(arr,np.eye(4,4))
-
+    
     # recast the estimation results
     nif = utils.recast_estimation_results(output, grid_parent)
     dat = nif.get_data()
-
+    
     # assert equivalence
     npt.assert_almost_equal(np.mean(dat[...,0]), x)
     npt.assert_almost_equal(np.mean(dat[...,1]), y)
     npt.assert_almost_equal(np.mean(dat[...,2]), sigma)
     npt.assert_almost_equal(np.mean(dat[...,3]), beta)
-    npt.assert_almost_equal(np.mean(dat[...,4]), hrf_delay)
-
+    
     # recast the estimation results - OVERLOADED
     nif = utils.recast_estimation_results(output, grid_parent, True)
     dat = nif.get_data()
-
+    
     # assert equivalence
     npt.assert_almost_equal(np.mean(dat[...,0]), np.arctan2(y,x),2)
[…]
