### 2016-11-29 7341786
TST: updated tests for new OG new HRF
