### 2016-11-29 98ff308
RF: made HRFs centered at 5s
