### 2016-11-30 #103 @arfon
> @arfon - how should I change the version number? (I assume not just editing the first post in this issue thread?) Apart from that, we're all set here!

😁 that's actually all it is. I'll do that now.
