### 2016-11-30 #103 @arfon
Sorry I didn't catch this sooner @kdesimone but could you move the references into a `paper.bib` file (and cite them from the Markdown)? We need them to be in a separate file for processing the submission.
