### 2016-11-30 58eff92
RF: moved refs to bib
