### 2016-11-30 024181c
ENH: added bib
