+++ popeye/paper.bib
@@ -1,0 +1,58 @@
+@online{popeye,
+  author = {Kevin DeSimone},
+  title = {popeye: a population receptive field estimation tool},
+  year = 2016,
+  url = {https://github.com/kdesimone/popeye},
+  urldate = {2016-11-10}
+}
+
+@Article{Dumoulin2008,
+author="Dumoulin SO
+and Wandell BA",
+title="Population receptive field estimates in human visual cortex",
+journal="NeuroImage",
+year="2008",
+volume="39",
+doi="10.1016/j.neuroimage.2007.09.034",
+url="http://dx.doi.org/10.1016/j.neuroimage.2007.09.034"
+}
+
+@Article{Thomas2014,
+author="Thomas JM,
+and Huber E,
+and Stecker E,
+and Boynton G,
+and Saenz M,
+and Fine I",
+title="Population receptive field estimates in human auditory cortex",
+journal="NeuroImage",
+year="2014",
+volume="109",
+doi="10.1016/j.neuroimage.2014.10.060",
+url="http://dx.doi.org/10.1016/j.neuroimage.2014.10.060"
+}
+
+@Article{Harvey2013,
+author="Harvey BM,
+and Klein BP,
+and Petridou N,
+and Dumoulin SO",
+title="Topographic organization of numerosity in the human parietal cortex",
+journal="Science",
+year="2013",
+volume="341",
+doi="10.1126/science.1239052",
+url="http://dx.doi.org/10.1126/science.1239052"
+}
+
+@Article{DeSimone2015,
+author="DeSimone K,
+and Viviano JA,
+and Schneider KA",
+title="Population receptive field estimation reveals two new maps in human subcortex",
+journal="Journal of Neuroscience",
+year="2015",
+volume="35",
+doi="10.1523/jneurosci.3840-14.2015",
+url="http://dx.doi.org/10.1523/jneurosci.3840-14.2015"
+}
