### 2016-11-30 100ff3f
RF: fixed refs inside paper
