### 2016-11-30 #103 @kdesimone
No problem, I added the paper.bib and moved refs out of paper.md.

https://github.com/kdesimone/popeye/blob/master/popeye/paper.bib
https://github.com/kdesimone/popeye/blob/master/popeye/paper.md
