### 2016-11-30 30f096e
ENH: trying to add dependencies to pypi install
