### 2016-11-30 #103 @arfon
> @arfon - is there as reason you set the archive DOI again?

This was because we were missing the field at the top of the issue (Archive: Pending) so the update that you requested didn't actually do anything :-)

I've now fixed this in https://github.com/openjournals/joss/commit/8fef34f8946c9d86e9486a7cc50cd2ddad485750
