### 2016-11-30 ff785b7
TST: added blanks to the sweeping bar
