### 2016-11-30 0baef68
ENH: added regularization functions to popeye for doing multivoxel fitting of a underspecified model parameter
