### 2016-11-30 ed68588
ENH: added a regularized CSS model for estimating one n per batch of voxels
