### 2016-12-01 4b9c7fd
ENH/RM: created 3 auditory models, removed the other two
