+++ popeye/auditory_baseline.py
@@ -9,12 +9,13 @@ warnings.simplefilter("ignore")
 import numpy as np
 from scipy.signal import fftconvolve
 from scipy.interpolate import interp1d
+from scipy.stats import linregress
 import nibabel
 
 from popeye.onetime import auto_attr
 import popeye.utilities as utils
 from popeye.base import PopulationModel, PopulationFit
-from popeye.spinach import generate_rf_timeseries_1D, binner
+from popeye.spinach import generate_rf_timeseries_1D
 
 class AuditoryModel(PopulationModel):
     
@@ -46,7 +47,7 @@ class AuditoryModel(PopulationModel):
         # invoke the base class
         PopulationModel.__init__(self, stimulus, hrf_model)
     
-    def generate_prediction(self, center_freq, sigma, beta, hrf_delay):
+    def generate_prediction(self, center_freq, sigma, beta, baseline, hrf_delay):
         
         r"""
         Generate a prediction for the 1D Gaussian model.
@@ -73,41 +74,46 @@ class AuditoryModel(PopulationModel):
         """
         
         # generate stimulus time-series
-        rf = self.receptive_field(center_freq, sigma)
-        
-        # if the tuning curve is running off the edges of the frequency space
-        if self.compare_rf_volumes(rf):
-            return np.inf
+        rf = np.exp(-((self.stimulus.freqs-center_freq)**2)/(2*sigma**2))
+        rf /= (sigma*np.sqrt(2*np.pi))
         
         # create mask for speed
         distance = self.stimulus.freqs - center_freq
         mask = np.zeros_like(distance, dtype='uint8')
-        mask[distance < (5*sigma)] = 1
+        mask[distance < (self.mask_size*sigma)] = 1
         
         # extract the response
-        response = generate_rf_timeseries_1D(self.stimulus.spectrogram, rf, mask)
-        
-        # bin the response to the resolution of the BOLD data
-        binned_response = self.binned_response(response)
-        
-        # roll it forward to correct for the binning
-        rolled_response = np.roll(binned_response, 0)
+        response = generate_rf_timeseries_1D(self.stimulus.stim_arr, rf, mask)
         
         # generate the HRF
         hrf = utils.double_gamma_hrf(hrf_delay, self.stimulus.tr_length)
         
         # pad and convolve
-        model = fftconvolve(rolled_response, hrf)[0:len(rolled_response)]
+        model = fftconvolve(response, hrf)[0:len(response)]
         
         # convert units
-        model = (model - np.mean(model))/np.mean(model)
+        # model = (model - np.mean(model)) / np.mean(model)
         
-        # scale by beta
+        # scale it
         model *= beta
+        
+        # offset
+        model += baseline
+        
+        # # # regress to find beta and baseline
+        # # p = linregress(model, self.data)
+        # 
+        # # save beta and intercept
+        # self.beta = p[0]
+        # self.intercept = p[1]
+        # 
+        # # scale it
+        # model *= self.beta
+        # model += self.intercept
         
         return model
     
-    def generate_ballpark_prediction(self, center_freq, sigma, beta, hrf_delay):
+    def generate_ballpark_prediction(self, center_freq, sigma, beta, baseline, hrf_delay):
         
         r"""
         Generate a prediction for the 1D Gaussian model.
@@ -135,60 +141,14 @@ class AuditoryModel(PopulationModel):
         
         """
         
-        return self.generate_prediction(center_freq, sigma, beta, hrf_delay)
-        
-    # seconds
-    @auto_attr
-    def length_in_seconds(self):
-        return np.ceil(self.stimulus.times[-1])
-    
-    # volumes
-    @auto_attr
[…]
