### 2016-12-01 0c07595
ENH/RM: created 3 auditory models, removed the other two
