### 2016-12-01 db68d5e
TST: shelving auditory tests for now
