### 2016-12-01 fd6477d
TST: shelving auditory tests for now
