### 2016-12-01 #103 @cMadan
@kdesimone, I think you're all set. @arfon needs to compile the revised paper.md and officially accept the submission.
