### 2016-12-01 #103 @arfon
Many thanks for the review @tknapen and for editing this submission @cMadan.

@kdesimone - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00103 🚀 🎉 💥
